Show move made by player in gamebord
Update bord based on specific symbol.

When the game is between players,they have their own symbol that will use to
play.And the move will be in the specific place of the board.This function
will add symbol to the specific place of the board.If there any no symbol
present.

The playerMove function will add the symbol where the user click,if the
box is already empty.and keep the symbol in gameboard array.By render the
gameboard we will able to see the move in the ui.